Add `hearing` flag to all entities
It seems rather wasteful to broadcast descriptive messages to entities
that probably don't have a player attached to them (rooms, items, and so
on), so, by default, entities are considered non-hearing, and any call
to entity.tell() will basically be a no-op.

This sort of entity state might be better as a component, actually...
but this works for now.