Remove "initialize" from World Foundry example
So, here's the problem.

I've found that describing a world layout (creating entities, linking
rooms together, etc.) using Python, etc. is extremely tedious, and I've
tried a couple of different methods of organizing the necessary "entity
bootstrapping" code to no avail.

At the end of the day, the goal of such a system is to emit a JSON file
of the starting entity set... and I'm starting to think it might be
easier to do this with a graphical editor of some kind. The issue with
that is (in lieu of some "Component schema" standard), it will
necessarily have knowledge of which Components exist, what data they
require, and so on. That is, such an editor would be specific to this
World Foundry example.

I'll be mulling over this. In the meantime, let's ditch the clunky
"initialize" script. Yes, this leaves the example without a way to
actually run it. For now.