Implement Stackable component
Now that it's possible to TDD these things, we can start implementing
some of the more complex Components and be reasonably sure that they'll
behave robustly.

That said, I'm sure there are still plenty of corner cases in this
Component, and this has revealed a few weak spots in the engine; see
the TODO comments.

(As an aside, the get/get_from/put_in/drop actions are crying out for
*some* kind of abstraction, though I haven't figured out what that is
yet.)