Beef up `prompt`, remove `command` and `listen`
With some ANSI escape code and nonportable readline tomfoolery, we can
remove the clunky "you gotta run two different processes" workflow.

This code is all on the chopping block anyway, since the message format
will be the purview of the world developers... but I have to admit it's
very useful for debugging.