Support structured messages
Messages sent to an entity were previously meant to be displayed
verbatim:

    > go north
    You're unable to do that.
    > go south
    Southern Forest
    Birdsong and the scent of pine needles fill the air.

    Contents:
        a frog
        a log

Now, they can be arbitrary JSON-serializable objects. This extends their
utility immensely: they can contain (or be) hints to the client like
colors or formatting, player status info (to be displayed in the
prompt), sounds, images, etc:

    > go north
    {
      "type": "message",
      "class": ["failure"],
      "text": "You're unable to do that."
    }
    {"type": "status", "hp": 100, "max_hp": 100}
    > go south
    {
      "type": "location",
      "title": "Southern Forest",
      "description": "Birdsong and the scent of pine needles...",
      "sound": {"path": "birds02.ogg", "loop": -1},
      "background": "forest.png"
    }
    {"type": "inventory", "items": ["a frog", "a log"]}
    {"type": "status", "hp": 100, "max_hp": 100}

It's up to the world developers and client developers to agree on a
protocol, of course.

A Renderer class allows world developers to control how these messages
are printed to the CLI in `figment prompt`. It's ugly, but it lets us
keep `figment prompt` around for a while longer.